requirements.txt
Babel==2.15.0
customtkinter==5.2.2
darkdetect==0.8.0
DateTime==5.5
packaging==24.1
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tk==0.1.0
tkcalendar==1.6.1
zope.interface==6.4.post2
